updateLab03Test.c
A simpler way to accomplish the same task. I uploaded both because I can't remember exactly what we knew at this point in class.

+++ students/severson.spring22/lab03Test.c/lab03Test.c.ino
@@ -5,7 +5,7 @@ int main(void) {
   DDRD = 0xff;
 
   while (1) {
-    for (int i = 0; i < 8; i++) {
+    for (uint8_t i = 0; i < 8; i++) {
       PORTD = (1 << i);
       _delay_ms(100);
     }
